+++ cs25-service/src/main/java/com/example/cs25service/domain/quiz/controller/QuizCategoryController.java
@@ -4,10 +4,13 @@ import com.example.cs25common.global.dto.ApiResponse;
 import com.example.cs25service.domain.quiz.dto.QuizCategoryRequestDto;
 import com.example.cs25service.domain.quiz.dto.QuizCategoryResponseDto;
 import com.example.cs25service.domain.quiz.service.QuizCategoryService;
+import com.example.cs25service.domain.security.dto.AuthUser;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.repository.query.Param;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -29,28 +32,4 @@ public class QuizCategoryController {
         return new ApiResponse<>(200, quizCategoryService.getParentQuizCategoryList());
     }
 
-    @PostMapping()
-    public ApiResponse<String> createQuizCategory(
-        @Valid @RequestBody QuizCategoryRequestDto request
-    ) {
-        quizCategoryService.createQuizCategory(request);
-        return new ApiResponse<>(200, "카테고리 등록 성공");
-    }
-
-    @PutMapping("/{quizCategoryId}")
-    public ApiResponse<QuizCategoryResponseDto> updateQuizCategory(
-        @Valid @RequestBody QuizCategoryRequestDto request,
-        @NotNull @PathVariable Long quizCategoryId
-    ) {
-        return new ApiResponse<>(200,
-            quizCategoryService.updateQuizCategory(quizCategoryId, request));
-    }
-
-    @DeleteMapping("/{quizCategoryId}")
-    public ApiResponse<String> deleteQuizCategory(
-        @NotNull @PathVariable Long quizCategoryId
-    ) {
-        quizCategoryService.deleteQuizCategory(quizCategoryId);
-        return new ApiResponse<>(200, "카테고리가 삭제되었습니다.");
-    }
 }
