+++ cs25-service/src/main/java/com/example/cs25service/domain/quiz/service/QuizCategoryService.java
@@ -21,29 +21,6 @@ public class QuizCategoryService {
 
     private final QuizCategoryRepository quizCategoryRepository;
 
-    @Transactional
-    public void createQuizCategory(QuizCategoryRequestDto request) {
-
-        quizCategoryRepository.findByCategoryType(request.getCategory())
-            .ifPresent(c -> {
-                throw new QuizException(QuizExceptionCode.QUIZ_CATEGORY_ALREADY_EXISTS_ERROR);
-            });
-
-        QuizCategory parent = null;
-        if (request.getParentId() != null) {
-            parent = quizCategoryRepository.findById(request.getParentId())
-                .orElseThrow(() ->
-                    new QuizException(QuizExceptionCode.PARENT_QUIZ_CATEGORY_NOT_FOUND_ERROR));
-        }
-
-        QuizCategory quizCategory = QuizCategory.builder()
-            .categoryType(request.getCategory())
-            .parent(parent)
-            .build();
-
-        quizCategoryRepository.save(quizCategory);
-    }
-
     @Transactional(readOnly = true)
     public List<String> getParentQuizCategoryList() {
         return quizCategoryRepository.findByParentIdIsNull() //대분류만 찾아오도록 변경
@@ -51,38 +28,4 @@ public class QuizCategoryService {
             ).toList();
     }
 
-    @Transactional
-    public QuizCategoryResponseDto updateQuizCategory(Long quizCategoryId, QuizCategoryRequestDto request) {
-        QuizCategory quizCategory = quizCategoryRepository.findByIdOrElseThrow(quizCategoryId);
-
-        if(request.getCategory() != null){
-            quizCategoryRepository.findByCategoryType(request.getCategory())
-                .filter(existingCategory -> !existingCategory.getId().equals(quizCategoryId))
-                .ifPresent(c -> {
-                    throw new QuizException(QuizExceptionCode.QUIZ_CATEGORY_ALREADY_EXISTS_ERROR);
-                });
-        }
-
-        quizCategory.setCategoryType(request.getCategory());
-
-        if(request.getParentId() != null){
-            QuizCategory parentQuizCategory = quizCategoryRepository.findByIdOrElseThrow(request.getParentId());
-            quizCategory.setParent(parentQuizCategory);
-        }
-
-        return QuizCategoryResponseDto.builder()
-            .main(quizCategory.getParent() != null
-                ? quizCategory.getParent().getCategoryType()
-                : null)
-            .sub(quizCategory.getCategoryType())
-            .build();
-    }
-
-    @Transactional
-    public void deleteQuizCategory(Long quizCategoryId){
-        if (!quizCategoryRepository.existsById(quizCategoryId)) {
-            throw new QuizException(QuizExceptionCode.QUIZ_CATEGORY_NOT_FOUND_ERROR);
-        }
-        quizCategoryRepository.deleteById(quizCategoryId);
-    }
 }
