+++ cs25-service/src/main/java/com/example/cs25service/domain/security/config/SecurityConfig.java
@@ -73,15 +73,8 @@ public class SecurityConfig {
 
                 //로그인이 필요한 서비스만 여기다가 추가하기 (permaiAll 은 패싱 ㄱㄱ)
                 .requestMatchers(HttpMethod.GET, "/users/**").hasAnyRole(PERMITTED_ROLES)
-                .requestMatchers(HttpMethod.POST, "/quizzes/upload/**")
-                .hasRole("ADMIN")//퀴즈 업로드 - 추후 ADMIN으로 변경
-
                 .requestMatchers(HttpMethod.POST, "/auth/**").hasAnyRole(PERMITTED_ROLES)
-                //.requestMatchers("/admin/**").hasRole("ADMIN")
-                .requestMatchers(HttpMethod.POST, "/quiz-categories/**").hasRole("ADMIN")
-                .requestMatchers(HttpMethod.PUT, "/quiz-categories/**").hasRole("ADMIN")
-                .requestMatchers(HttpMethod.DELETE, "/quiz-categories/**").hasRole("ADMIN")
-                .requestMatchers(HttpMethod.DELETE, "/quizzes/**").hasRole("ADMIN")
+                .requestMatchers("/admin/**").hasRole("ADMIN")
                 .requestMatchers(HttpMethod.POST, "/crawlers/github/**").hasRole("ADMIN")
 
                 .anyRequest().permitAll()
